src/test/java/WagenparkbeheerTest.java
import static org.junit.Assert.*;

import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;

public class WagenparkbeheerTest {

	@BeforeClass
	public static void setUpBeforeClass() throws Exception {
	}

	@AfterClass
	public static void tearDownAfterClass() throws Exception {
	}

	@Before
	public void setUp() throws Exception {
	}

	@After
	public void tearDown() throws Exception {
	}

	@Test
	public void testGetData() {
		Wagenparkbeheer w1 = new Wagenparkbeheer();
		assertEquals(0, w1.val);
		w1.getData();
		assertEquals(10, w1.val);
	}

}
